% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEUDataRd.R
\docType{data}
\name{CEUData}
\alias{CEUData}
\alias{CEUSNP}
\alias{CEUDist}
\alias{hapmapCEU}
\title{Example data set for LDheatmap}
\format{CEUSNP: A dataframe of SNP genotypes.
Each row represents an individual.
Each column represents a SNP.

CEUDist: A vector of integers, representing SNP physical map locations on the chromosome.}
\source{
International HapMap Project \url{www.hapmap.org}
}
\usage{
data(CEUData)
}
\description{
CEUSNP:  Genotypes on 15 SNPs for 60 people

CEUDist: Physical map positions of the 15 SNPs in CEUSNP
}
\details{
Data on SNPs with minor allele frequency greater
than 5\% from a 9kb region of chromosome 7 (base positions 126273659
                                           through 126282556 from release 7 of the International HapMap Project).
Genotypes from 30 parent-offspring trios (both
                                         parents, one offspring) were obtained.
The 30 trios are taken from the so-called CEPH families, a set of
multi-generational families from Utah with ancestry from northern and
western Europe. From this set of 90 people, 60 parents were extracted.
}
\examples{
data(CEUData)
}
\references{
The International HapMap Consortium. A haplotype map of
the human genome. Nature 437, 1299-1320. 2005.
}
\keyword{datasets}
