% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertVCF.R
\name{convertVCF}
\alias{convertVCF}
\title{Convert a VCF file into genotype data}
\usage{
convertVCF(file, phased = NULL, subjects = NULL, ...)
}
\arguments{
\item{phased}{If \code{TRUE} the output genotype data are in the form of a XSnpMatrix object. Otherwise, they are in the form of a SnpMatrix object. 
If it is unspecified, the phasing status will be determined by checking the first entry in the GT section of the VCF file.}

\item{subjects}{A character or factor containing subject IDs. If supplied, genotype info of only those subjects will be returned. 
This should be a subset of the sample IDs in the VCF file.}

\item{...}{Additional arguments to be passed to vcfR::read.vcfR().}

\item{file}{A filename for a variant call format (vcf) file.}
}
\value{
A list which contains the following components:
\item{genetic.distances}{ A numeric vector of the reference positions of SNPs. }
\item{subjectID}{ A character vector of IDs of the subjects which the returned genotype data belong to. }
\item{data}{An object of SnpMatrix/XSnpMatrix class from the 'snpStats' package containing genotype data. }
}
\description{
Convert a VCF file into a list composed of genetic distances, subject IDs, and a SnpMatrix/XSnpMatrix object.
}
\seealso{
\code{\link[vcfR]{read.vcfR}}, \code{\link[snpStats]{SnpMatrix-class}}, \code{\link[snpStats]{XSnpMatrix-class}}
}
\keyword{hplot}
