\name{LDheatmap-internal}
\title{Internal LDheatmap functions}
\alias{backbone}
\alias{jiggle}
\alias{myRainbow.colors}
\alias{makeImageRect}
\alias{LDheatmap.Legend.add}
\alias{LDheatmap.Map.add}
\description{Internal LDheatmap functions.}
\details{These are not to be called by the user or are undocumented.}
\keyword{internal}

