\name{LDheatmap-internal}
\title{Internal LDheatmap functions and data files}
\alias{preDrawDetails.ldheatmap}
\alias{postDrawDetails.ldheatmap}
\alias{preDrawDetails.symbols}
\alias{postDrawDetails.symbols}
\alias{constructVP}
\alias{drawLDheatmapGrob}
\alias{moveTitles}
\description{Internal LDheatmap functions.}
\details{These are not to be called by the user or are undocumented.}
\keyword{internal}
